04_demo/app/backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
python-multipart>=0.0.9,<1
pydantic>=2.9,<3
pydantic-settings>=2.5,<3
httpx>=0.27,<1
python-dotenv>=1,<2
langchain>=0.3,<0.4
langchain-openai>=0.2,<0.4
langchain-community>=0.3,<0.4
langchain-text-splitters>=0.3,<0.4
pypdf>=5,<6
pytest>=8,<9
pytest-asyncio>=0.24,<1
orjson>=3.10,<4
SQLAlchemy>=2.0,<3
psycopg[binary]>=3.2,<4
reportlab>=4.2,<5
PyJWT>=2.9,<3
argon2-cffi>=23.1,<24
